# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
      "AbstentionReason": {
        "enum": [
          "thin_local_record",
          "dominated_by_pooling",
          "interval_too_wide",
          "stale_jurisdiction_data",
          "unresolved_jurisdiction_chain",
          "degenerate_training_corpus"
        ],
        "title": "AbstentionReason",
        "type": "string"
      },
...
      "Alternative": {
        "additionalProperties": false,
        "properties": {
          "abstained": {
            "default": false,
            "title": "Abstained",
            "type": "boolean"
          },
          "approval_probability": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Probability"
          },
          "by_right": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "By Right"
          },
          "distance_km": {
            "minimum": 0.0,
            "title": "Distance Km",
            "type": "number"
          },
          "expected_value_rank": {
            "title": "Expected Value Rank",
            "type": "number"
          },
          "jurisdiction": {
            "title": "Jurisdiction",
            "type": "string"
          },
          "jurisdiction_slug": {
            "title": "Jurisdiction Slug",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "jurisdiction",
          "jurisdiction_slug",
          "distance_km",
          "expected_value_rank"
        ],
        "title": "Alternative",
        "type": "object"
      },
...
      "Determination": {
        "additionalProperties": false,
        "properties": {
          "abstained": {
            "default": false,
            "title": "Abstained",
            "type": "boolean"
          },
          "abstention_reasons": {
            "items": {
              "$ref": "#/components/schemas/AbstentionReason"
            },
            "title": "Abstention Reasons",
            "type": "array"
          },
          "approval_probability": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Probability"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "credible_interval_80": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "maximum": 1.0,
                    "minimum": 0.0,
                    "type": "number"
                  },
                  {
                    "maximum": 1.0,
                    "minimum": 0.0,
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credible Interval 80"
          },
          "interval_kind": {
            "anyOf": [
              {
                "enum": [
                  "credible",
                  "bootstrap"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Kind"
          },
          "local_base_rate": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "The jurisdiction's own historical rate for this use class. Rendered as a dashed marker against the interval, because that single comparison is the most informative thing on the page.",
            "title": "Local Base Rate"
          },
          "probability_of_rule_change_before_decision": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability Of Rule Change Before Decision"
          },
          "time_to_decision_months": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeToDecision"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "Determination",
        "type": "object"
      },
      "Driver": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "enum": [
              "positive",
              "negative",
              "neutral"
            ],
            "title": "Direction",
            "type": "string"
          },
          "evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null only for a driver derived entirely from structured registry data, such as the election calendar, where there is no quotable document.",
            "title": "Evidence Id"
          },
          "factor": {
            "title": "Factor",
            "type": "string"
          },
          "group": {
            "title": "Group",
            "type": "string"
          },
          "plain_language": {
            "minLength": 10,
            "title": "Plain Language",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "weight": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "factor",
          "group",
          "direction",
          "weight",
          "plain_language"
        ],
        "title": "Driver",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Mitigation": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "minLength": 10,
            "title": "Action",
            "type": "string"
          },
          "basis": {
            "minLength": 10,
            "title": "Basis",
            "type": "string"
          },
          "expected_delta": {
            "description": "Change in approval probability if the action is taken. Computed by re-scoring with the feature changed, never by asking a language model to estimate it.",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Expected Delta",
            "type": "number"
          }
        },
        "required": [
          "action",
          "expected_delta",
          "basis"
        ],
        "title": "Mitigation",
        "type": "object"
      },
...
      "PortfolioRequest": {
        "additionalProperties": false,
        "description": "Section 5.4 product 2, the wedge feature. Up to 500 sites in, ranked out.",
        "properties": {
          "sites": {
            "items": {
              "$ref": "#/components/schemas/ScoreRequest"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Sites",
            "type": "array"
          }
        },
        "required": [
          "sites"
        ],
        "title": "PortfolioRequest",
        "type": "object"
      },
      "PortfolioResponse": {
        "properties": {
          "abstained": {
            "description": "How many we refused to put a number on.",
            "title": "Abstained",
            "type": "integer"
          },
          "note": {
            "default": "Ranked by approval probability. Abstentions sort last and carry no number, because a ranked list that quietly treats an abstention as a low score would make refusing to answer indistinguishable from answering badly.",
            "title": "Note",
            "type": "string"
          },
          "ranked": {
            "items": {
              "$ref": "#/components/schemas/PortfolioRow"
            },
            "title": "Ranked",
            "type": "array"
          },
          "scored": {
            "description": "How many carry a probability. Excludes abstentions, so scored plus abstained equals submitted.",
            "title": "Scored",
            "type": "integer"
          },
          "submitted": {
            "description": "How many sites were in the request.",
            "title": "Submitted",
            "type": "integer"
          }
        },
        "required": [
          "ranked",
          "submitted",
          "scored",
          "abstained"
        ],
        "title": "PortfolioResponse",
        "type": "object"
      },
      "PortfolioRow": {
        "properties": {
          "abstained": {
            "title": "Abstained",
            "type": "boolean"
          },
          "approval_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Probability"
          },
          "credible_interval_80": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credible Interval 80"
          },
          "data_depth": {
            "title": "Data Depth",
            "type": "integer"
          },
          "jurisdiction": {
            "title": "Jurisdiction",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "months_p50": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Months P50"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "rule_change_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Change Probability"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "label",
          "jurisdiction",
          "approval_probability",
          "credible_interval_80",
          "abstained",
          "months_p50",
          "rule_change_probability",
          "data_depth",
          "stale",
          "public_id"
        ],
        "title": "PortfolioRow",
        "type": "object"
      },
...
      "Relief": {
        "description": "The instrument being requested. A project usually needs several.",
        "enum": [
          "rezoning",
          "special_use_permit",
          "conditional_use_permit",
          "variance",
          "site_plan_approval",
          "comprehensive_plan_amendment",
          "text_amendment",
          "annexation",
          "development_agreement",
          "tax_abatement",
          "subdivision_plat",
          "other"
        ],
        "title": "Relief",
        "type": "string"
      },
      "Score": {
        "additionalProperties": false,
        "description": "What a customer receives. Appendix 19.3.",
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/Alternative"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "determination": {
            "$ref": "#/components/schemas/Determination"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/Driver"
            },
            "title": "Drivers",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "features_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Features Hash",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "mitigations": {
            "items": {
              "$ref": "#/components/schemas/Mitigation"
            },
            "title": "Mitigations",
            "type": "array"
          },
          "precedents": {
            "items": {
              "$ref": "#/components/schemas/Precedent"
            },
            "title": "Precedents",
            "type": "array"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "site": {
            "$ref": "#/components/schemas/Site"
          }
        },
        "required": [
          "public_id",
          "generated_at",
          "site",
          "determination",
          "provenance",
          "features_hash"
        ],
        "title": "Score",
        "type": "object"
      },
      "ScoreRequest": {
        "additionalProperties": false,
        "description": "One site. Either a coordinate pair or a jurisdiction slug, and a use class.",
        "properties": {
          "acres": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acres"
          },
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Score as the world was known on this date. Used for back testing, not for a live quote. Omitting it means today.",
            "title": "As Of"
          },
          "by_right": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "By Right"
          },
          "capacity_mw": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Mw"
          },
          "include_alternatives": {
            "default": true,
            "title": "Include Alternatives",
            "type": "boolean"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 8,
                "pattern": "^[a-z]{2}-[a-z]{2}-[a-z0-9-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Registry slug, e.g. us-va-loudoun",
            "title": "Jurisdiction"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "parcel_ids": {
            "items": {
              "maxLength": 64,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Parcel Ids",
            "type": "array"
          },
          "relief_sought": {
            "items": {
              "$ref": "#/components/schemas/Relief"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Relief Sought",
            "type": "array"
          },
          "use_class": {
            "$ref": "#/components/schemas/UseClass"
          }
        },
        "required": [
          "use_class",
          "relief_sought"
        ],
        "title": "ScoreRequest",
        "type": "object"
      },
...
      "UseClass": {
        "enum": [
          "data_center_hyperscale",
          "data_center_colocation",
          "data_center_edge",
          "solar_utility",
          "wind_onshore",
          "battery_storage",
          "substation_transmission",
          "industrial_general",
          "warehouse_logistics",
          "residential_multifamily",
          "other"
        ],
        "title": "UseClass",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Calibrated permission risk forecasts with a published accuracy record.\n\nEvery response carries an interval and a data date. When the evidence is too thin the response abstains, and an abstention is a successful response rather than an error. The accuracy record and the jurisdiction profiles are public and need no key.",
    "summary": "A rating bureau for the right to build.",
    "title": "Permission Bureau",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/portfolio": {
      "post": {
        "description": "Score up to 500 sites and rank them.\n\nThis is the wedge feature. Section 9.2: never compete on one site, where a good local lawyer is\noften better than any model. Compete on three hundred, which no lawyer can do at any price.",
        "operationId": "score_portfolio_v1_portfolio_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Screen a portfolio",
        "tags": [
          "score"
        ]
      }
    },
...
    "/v1/score": {
      "post": {
        "description": "Probability of approval, a time distribution, drivers with evidence, and alternatives.\n\nReturns 200 with ``abstained`` set to true when the evidence is too thin for a number. An abstention\nis a successful response, not an error: the customer asked a question and got the honest answer.",
        "operationId": "score_one_v1_score_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Score"
                }
              }
            },
            "description": "The full score object, or an abstention."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score one site",
        "tags": [
          "score"
        ]
      }
    },
    "/v1/score/as-of": {
      "get": {
        "description": "Every score carries ``data_as_of``. This exposes it without running a score.",
        "operationId": "data_as_of_v1_score_as_of_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Data As Of V1 Score As Of Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The date the served data is current to",
        "tags": [
          "score"
        ]
      }
    },
    "/v1/score/{public_id}": {
      "get": {
        "description": "Read back a stored prediction by its public identifier.\n\nRetrieval rather than recomputation, deliberately. A score is a dated statement about what was known on a\nparticular day, and a report opened next month has to show the same number the memo in the deal file\nshows. Recomputing on read would quietly revise a published prediction, which section 8.9 forbids.",
        "operationId": "get_score_v1_score__public_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "title": "Public Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Score"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve a score that was already produced",
        "tags": [
          "score"
        ]
      }
    },
...
  }
}